Add Reference Material
This adds a reference material section to the left side navigation, including Teams links and phone tree information (some of which was originally included in Andrew's script doc), plus a link to the OneNote page for finding serial numbers and a military phonetic alphabet link for spelling over the phone. These references are incomplete and mostly proof-of-concept. This commit also includes a fix for an export bug where the Additional Information was disconnected from the RAFTA section and cleans up a bit of the formatting.