Proaim Mail Template Fix
This fixes issue #24 which was not working because the '&' in line 17 (proaimpmemailgen) and line 15 (proaimserviceemail) should have been a '?' to indicate parameters list begin. As part of the troubleshooting for this issue, I consolidated these into a format matching the prevmaintemail.js file (setting subject and body vars, encoding as a separate action, constructing mailto link separately, etc).