FSE PO Request Clean-up
This change cleans up some of the verbiage on the FSE request template and includes some processing of the instrument and serial number fields on the form to make the template a little more context-aware to require less manual tweaking before hitting send.

Commit also includes cleanup of the dynamicfields file to remove old, commented code.