SNAP Library
This commit lays the groundwork for some of the suggestions highlighted in issue #44 to simplify and unify some of the code. It doesn't do much in this version, mostly creates a dynamically populated copyright string and produces a console output start-up message.

But, the re-use of the `curr_date` in exportAll is an example of unifying the codebase, and the version string should be a single input point for setting the active or working version to avoid confusion.