Fix syncFields Error
Removal of the device-serial-number duplicate field in the HTML caused the syncfields script to throw an error into the console. It appears pretty benign, not causing any other functionality to fail, but this fixes the issue. I do wonder if we need the dupe field on description/problem-description though if it's the only one, although the gray field highlighting makes it clear why it appears in more than one place (as it's included in the Request section and RAFTA notes).