Forum Independent Toggle
This change primarily adds a new checkbox to the Remote Support section which toggles the FORUM checklist fields on/off, allowing FORUM troubleshooting details to be included regardless of instrument selection.

This change also removes the patch number from the README file so it doesn't need to be updated with each patch.