Select Dependencies
This change, while significant in terms of modified files and lines, amounts to a couple of relatively small QoL improvements.

Selectable Error Code Fields - Rather than text input fields for the error group, error code, and action code fields, these have been changed to select boxes with options ported over from CRM.

Field Dependencies - Select fields are designated as parent fields for dynamically-populated dependent fields which apply their option lists based on the context of the parent. In all cases, the default for the dependent field is to be disabled and only enabled when a non-default option for the parent is set. The new dependencies are:
 - Billing Type :: Service Contract
 - Request Came From :: Request Source
 - Error Group :: Error Code

Dbugmsg - a new developer function correlated to a new debug_level setting which allows developers to more easily add debug logging messages to the code and provides finer-grained control of how much logging is sent to the console.