Fix Export Date
This is a fix for issue #80.

RCA: Underlying issue appears to be that the reset button did not modify the constant curr_date which was being used in the exportAll script. I didn't notice this because I spend a lot of time with SNAP using the console log for debugging so I was refreshing the whole page (which reloads the snaplib file and therefore refreshes the date).

This change creates a new startup routine for the console so a refresh isn't necessary (it actively clears the console.log) and changes the curr_date variable so it's not a constant but is refreshed on every "reset" press. This also moves the setting variables from constants since I suspect that would cause issues when we get around to addressing enhancement #59).