v3.0.37
This fairly extensive update addresses three of the most annoying long-standing issues:

Issue #10 - Improves the overall UI by making the default page much smaller and adds fields based on the context of the selections as users work down the page. This also allows arbitrary fields to be toggled on/off by specific grouping (in this case, instrument or page functionality based).

Issue #76 - Technically a workaround to the CSS issues that plagued the previous dynamic fields functionality, the above fixes for Issue #10 allow the regular two-column format to persist on dynamic fields.

Issue #51 - Includes many of the proposed UI improvements including a persistent navigation bar on the top of the page, Remote Service, review stations, and FORUM inclusion by toggle, specific field for network configuration, and a more natural-feeling page flow.