Sandbox Serial Tooltip POC
This change includes a test/sample of a potential enhancement leveraging the tooltip text used in the Shipping Address field to provide some contextual information about the serial number: format, locations, etc. But this would be context-sensitive based on the selection from the Instrument field.

This isn't fully implemented because there's no open issue for it and I'm not sure yet about the format/utility (all the information would basically be copied from the OneNote doc that is already linked under Reference Material), but thought a proof-of-concept would be easier to discuss than trying to describe it.