Bugfix
Fixes a minor bug in the dynamicfields.js code where an error would generate because of the serial number tooltip POC. This fix restricts the feature to sandbox.

Also includes a few small adjustments to the tooltip styles and a fix for an issue with the tooltip before an instrument is selected.