Bugfix process_disk_space
This change fixes a bug in exportAll where the process_disk_space function fix from an earlier commit was not carried over from printintnotes to the other file. To resolve this, the duplicated function definition has been moved into snaplib and tested to verify it works in both the Copy Internal Notes routine as well as Export Interaction.