Template POC
openEmailTemplate (Feature Improvement)
This is mostly a POC update. It addresses (primarily) issue #44 in that it presents a mechanism for unifying the ever-growing list of email templates into a single file and re-usable functions. In the process, this also lays the groundwork to tackle other issues such as #33, #58, and #100 by providing the templates object which can be instantiated to provide access to template literals and text blobs without the need to hunt them down across various files and functions.

This is not a wholesale migration, however. It is presented here as a proof of concept limited to the Sandbox for testing and collaborative evaluation. Even then, this functionality has been applied to only about a third of the total email templates (and no other text-based features like CCT Copy functions). The reason for this is largely in that, while the openEmailTemplate implementation does a lot in terms of re-using code, eliminating unnecessary re-declaration of variables, and separating presentation from logic, it does add some complexity. In particular, the template's context element is very loosely structured and could be prone to bugs if not handled cautiously.

The templates object definition (templates.js) is also liable to become quite large. I doubt for the use scenario of SNAP it represents a significant performance hit, but it's worth being aware of the possibility. That said, after implementing six templates (including a brand-new one for Smart Services), this approach has reduced total lines of code from about 325 across seven files to just over 150 lines in three files so it is in the neighborhood of 50% more efficient from that perspective.

Other Included Changes
- Fixes a bug in the device matrix and endOfSupport functionality preventing HFA models from being handled correctly
- [Sandbox only] Re-organizes the Processes navigation list to be alphabetical (except Misc Email Templates which remains at the bottom for clarity)
- [Sandbox only] openEmailTemplate implementation of Parts Request template fixes a bug in the handling of "Free of Charge" overlay form element
- [Sandbox only] Adds a copy button to the Billing Contact dynamic form elements which copies the deferred billing string to the clipboard
- [Sandbox only] Not a committed change, but sandbox_darkmode.css contains some commented lines which were meant to address issue #40 (closed as wontfix). It almost works, just need a way to lock the sidebar div into the viewing frame (similarly to how the top navigation bar works) otherwise scrolling the main page moves the background of the sidebar. Left here to iterate on later.