Bugfixes
This fixes a major bug in 3.1.2 where the template would not load (issue #122) and a minor bug where the inactive RAFTA reported issue field was not correctly synchronizing when the Call Scenario option was used (issue #123).
This also moves the code for the description field sync-up to formsupport.js and removes the syncFields.js file.