Stub references (at least) for all devices
This adds index references for all devices in the (current) list. The productData model is scheduled to have a brainstorming/vetting pass on Thursday to clean up some of its messier elements and any changes resulting from that will need to be reflected in the reference doc structure and presentation.