Sample Overviews
The overview docs included in product references have been templatized here. These are mostly stubs as the reference functionality is still in its earliest incarnation, but this approach prevents most of the broken links and is sufficient for early adopter testing of the feature.

Also includes some logging clean-up on the loadXml functions.