Bugfixes
This change fixes issue #144.

It also updates the template for the customer notes to remove FSS information from the output text, and updates how billing request templates handle the SLA data (now that it's configurable in the user settings). This required adding a new library file, numbers2words, which is now available to all functions loaded after openEmailTemplate.