Add Customer Ticket Summary
This change adds a new function to send a customer ticket summary via email, replacing similar functionality in CRM (though that one is broken and not configurable without escalation to Germany).

The template verbiage has not yet been fully vetted, but committing as-is as there is no ETA on management approval and the template can always be updated at a later time.

This change also includes a bugfix for the manualOpenEmailTemplate, particularly the Chrome mailto link limit where the process box wasn't displaying the correct recipient addresses and the formatting has been improved.

A few more instances of the deprecated 'debugmsg' function have been replaced as well.