Add Time Zone Info
This change adds a hidden Time Zone field (for reference by other functions and features) as well as a text field on the Scheduled Callback overlay.

Both are keyed off a change in the Instrument/Shipping Address (id: instrument-address) field, which uses the ZIP code to identify the TZ.

Still needs a bit more testing before merging, but looks ok so far.

Also includes minor bugfixes and template tweaks.