Simplify SLA Message
This change attempts to fix the SLA scheduling estimation by simplifying away from specific times/deadlines and moving to a more "end of day/tomorrow/next week" style.

Also included is a fix for the new SNAP form link which was not working as intended after adding the Customer/Instrument options.