Case History
This change includes a first pass at a (semi-)hidden history field. This allows for some automated tracking of progress on a given call/issue, particularly for users who leverage the "Save as XML" functionality to track ongoing incidents.

As of this commit, only a handful of events update the case history. Additional events can be added in the future via the logHistoryEvent function.