Form Elements and XML Load Bug
This change addresses some issues with the XML loading, particularly related to form fields outside the form element with id "snap-form". This form was repositioned to address a visual bug (issue #171) but regresses the XML load feature for overlay form groupings.

The solution here is to add a second form for those elements and update the readFile sequence to manage multiple forms.

Also includes an update to one of the time zones in the TZ data object and an improvement to the logging for the history write function.