Add KB/Solutions Feature
Adds framework for the Solutions feature, which compares Instrument/Model, Software Version, and Problem/Issue Description field against a database of common issues with known solutions and presents the user with a weighted list of possible fixes to the customer issue.

This update also includes some CSS changes for the light mode which makes the Common Actions menu usable in that mode.

TODO: Expand solutions list and add feature to auto-populate some of the form fields from the solution card.