Product Data Update, EoS Process
This change includes an update to the productData object, adding a guaranteed support schema item and applying new end of support data. Also included is a feature to open the process box whenever an EoS or EoGS device is selected with relevant data regarding handling support for products that have reached end of support.