Sync all stylesheets, add common parts
This change attempts to synchronize all updates to the main.css merged as part of commit/merge ca87a33 into the sandbox and darkmode stylesheets. Formatting changes and minor tweaks abound.

This also adds a function similar to the common actions to the Spare Parts overlay for commonly requested parts.