Fix anchor link page-hop
A number of code-managed link elements include onclick or listener-based event handling and managed the default behavior by setting the href to the closest relevant anchor reference. This resulted in the page focus shifting around when interacting with various elements that did not imply page relocation.

This change modifies all the relevant functions and onclick calls to override the event (without using the deprecated event.preventDefault() call).

Also includes some fixes for relative pathing in templates and typos.